type conversion, metadata number values are always double-s
git-svn-id: https://flazr.svn.sourceforge.net/svnroot/flazr/trunk/flazr@66 5dee199b-8608-4927-992f-dc734163a847

+++ src/main/java/com/flazr/rtmp/message/Metadata.java
@@ -180,8 +180,8 @@ public abstract class Metadata extends AbstractMessage {
             );
             VideoSD video = movie.getVideoSampleDescription();
             map(map,
-                pair("width", video.getWidth()),
-                pair("height", video.getHeight()),
+                pair("width", (double) video.getWidth()),
+                pair("height", (double) video.getHeight()),
                 pair("videocodecid", sampleType)
             );
         }
